Parse escaped backslashes correctly
Currently CSS rules containing escaped backslashes inside a string will fail
to be parsed correctly.  A minimal example of a failing CSS rule one which prepends
a backslash to an element using the '::before' pseudo-element:

a::before {
    content: "\\"
}

This will currently fail with the following message:

Error: Unmatched " (line x, char y) (line x)

The fix is to check for this condition in the do-while loops that are part of
the string parsing logic.